Variabilize behaviors "building cyrus from source" (cyrus upstream) vs. "testing distribution package" (Debian)
By default, and in order to be backward-compatible (so upstream doesn't
have to modify any of its existing CI calls), functions.sh and
testscript.sh will continue building cyrus from source and testing that.

If the BUILD_CYRUS_FROM_SOURCE variable is set to something else than
"yes", however, the Debian package will be used instead; this is what
the maintainers of cyrus-imapd in Debian use for autopkgtest.

A second extra environment variable, CASSANDANEOPTS, is also added: it
is set to "Cyrus::ImapTest" in Debian, in order to limit the range of
tests being run. Its default value is "", once again to keep the
behavior backward-compatible with the previous implementation.