Debian/functions.sh: set up cassandane.ini earlier
Part of `make` expects to be able to read cassandane.ini to find
the Cyrus perl modules, in order to properly syntax check annotator.pl.
So set it up before running make.

We got away with this in the past because the paths happened to
be the same as the defaults that would be assumed if cassandane.ini
was missing, but cassandane.ini not being found is now an error
(as of 53a584c580345643187b3e6cc9a7640121a4a2a2).