cyruslibs: clean up after install
The working directory used for cyruslibs is a good 2 GB.  No reason to
keep that in the final image, so remove it before the step finishes.

This also converts to a shallow clone.

This **also** uses the cyruslibs-fastmail-v58 tag, because know that
master is not building properly now. This will be fixed.