combine more get-and-build steps into single RUNs
This creates fewer layers and will make further post-build cleanup
possible.

Note also that this eliminates "groupadd -r mail", which would always
fail on Debian, where the mail group will exist by now.  If we're
concerned this might not happen, we can do something like "try to run it
and check that $? is 0 or 9", as 9 means "name already in use".  I think
we can skip that, though.