cyd build: Add --ubsan-trap mode to bail when errors are detected
By default, asan aborts if it detects an error, but ubsan doesn't.

This means asan checks will fail cunit tests, but ubsan checks won't.

I've added --ubsan-trap which makes it so ubsan behaves like asan so we
can get a CI build of at least make check with the sanitizer.