GitHub Action: build an ARM image, add "dev" branch
Most importantly, this commit switches the workflow to a matrix that
will build both a "normal" (amd64) container and now also an ARM
container.  While this might have multiple uses, I'm using it to do
development on a VM on my MacBook M2, at least when disconnected from
the internet and my usual Linux build host.

This commit removes mention of the "main" branch, to which we may yet
rename master.  Instead, it adds a "dev" branch, which will generate
"dev" labels for images.  This, because we don't build containers from
pull requests.  When testing automation, the dev branch provides a
convenient way to test.