Dockerfile: update versions of Dovecot and imaptest
We stick to a known-working version of Dovecot and imaptest, but that
version stopped building recently.  During its doc-making step, it wants
to fetch an archive of man pages from the internet.  The URL used has
become a cyclical redirect, and it's not used by real builds anymore, so
pushing for a fix seems churlish.

Instead, this just moves us to a very recent Dovecot and imaptest
commit.

Note also that it no longer requires that we use our own fork of
imaptest, which was just "imaptest pinned to an old version".  This
means we can stop maintaining one more little thing!