dar: after prune, wait for autoremoval to complete
By default, dar will create containers that are autoremove, meaning that
when stopped, they implicitly rm themselves.  This removal is done in
the background, so that the `stop` finishes immediately, but the `rm` is
still happening.  Often this is very fast.  Sometimes, it isn't.

That's a problem, because `dar prune && dar start` will crash when
`start` finds that the old container is still running, using the same
name that we want to reuse.

Let's wait, instead.  We'll just run `docker inspect` over and over
until it fails -- and we'll assume, *probably* safely, that it's failing
because there is no longer a container with the given id.